pyproject.toml
[tool.poetry]
name = "authdemo"
version = "0.1.0"
description = ""
authors = ["balancy <balancy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
setuptools = "^57.0.0"
python-multipart = "^0.0.5"
aiofiles = "^0.7.0"
Jinja2 = "^3.0.1"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
